Moved the Contract class as AbstractContractAspect
